% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\docType{data}
\name{coords}
\alias{coords}
\title{Coordinate of region in South Sumatera}
\format{
A \code{tibble} with 17 rows as Region/City and 2 column,which are:
\describe{
\item{"Longitude"}{longitude coordinate for each location}
\item{"Latitude"}{latitude coordinate for each location}
}
}
\source{
https://www.bps.go.id/
}
\usage{
coords
}
\description{
A simple \code{tibble} dataset containing the coordinate region In South.
}
\keyword{datasets}
