% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedata.R
\docType{data}
\name{simulatedata}
\alias{simulatedata}
\title{Sample Data for simulate analysis data}
\format{
A \code{tibble} with 100 observation time and 17 column as location,which are:
\describe{
\item{"PALEMBANG"}{a value dataset for PALEMBANG}
\item{"LUBUKLINGGAU"}{a value dataset for LUBUKLINGGAU}
\item{"OGAN KOMERING ULU SELATAN"}{a value dataset for OGAN KOMERING ULU SELATAN}
\item{"OGAN KOMERING ULU"}{a value dataset for OGAN KOMERING ULU}
\item{"OGAN KOMERING ILIR"}{a value dataset for OGAN KOMERING ILIR}
\item{"MUSI RAWAS"}{a value dataset for MUSI RAWAS}
\item{"OGAN ILIR"}{a value dataset for OGAN ILIR}
\item{"PAGAR ALAM"}{a value dataset for PAGAR ALAM}
\item{"BANYU ASIN"}{a value dataset for BANYU ASIN}
\item{"OGAN KOMERING ULU TIMUR"}{a value dataset for OGAN KOMERING ULU TIMUR}
\item{"EMPAT LAWANG"}{a value dataset for EMPAT LAWANG}
\item{"PRABUMULIH"}{a value dataset for EMPAT LAWANG}
\item{"LAHAT"}{a value dataset for  LAHAT}
\item{"MUSI RAWAS UTARA"}{a value dataset for  MUSI RAWAS UTARA}
\item{"PENUKAL ABAB LEMATANG ILIR"}{a value dataset for  PENUKAL ABAB LEMATANG ILIR}
\item{"MUARA ENIM"}{a value dataset for  MUARA ENIM}
\item{"MUSI BANYUASIN"}{a value dataset for  MUSI BANYUASIN}
}
}
\usage{
simulatedata
}
\description{
A simple \code{tibble} that is generate from random normal
distribution for simulate seasonal generalized space-time autoregressive model.
}
\keyword{datasets}
